Test screen (#38)
* Display: Text knows how to break into a new line

This modifies the top offset by font height +2 px if a \n is parsed. It also resets the column.

* Display: Text knows carriage return and tab

\r moves the rendering back to the column origin.
\t jumps to the next divisible by 8 column.

* App: Read battery voltage at startup

This causes the battery voltage to be known when screen is rendered initially

* GPS: Add 1 hour to time

Don't know why but i could not get the DST to work. Maybe have a look when it is winter :^)

* GUI: Reset display for longer

This might help with the display not beeing enabled after enabeling power to it.

* GUI: Expose battery level value

* GUI: Only create the test screen once

* Graph: fill the background with white

* Graph: allow the drawing of more data than pixels

We have floats as x and y step so we can allow drawing of more data than pixels. The data will overlap though.

* Testscreen: Display information about GPS and Tasks

* Test Screen: Show ADC values in Graph and current value in numbers